Rename Ciro boundary file
