+++ scripts/fetchCiroBoundary.mjs
@@ -60,8 +60,8 @@ async function main() {
 
   const featureCollection = { type: 'FeatureCollection', features: [{ type: 'Feature', properties: {}, geometry: geo }] };
   await fs.mkdir('public/geo', { recursive: true });
-  await fs.writeFile('public/geo/ciro_marina.geojson', JSON.stringify(featureCollection, null, 2));
-  console.log('Saved to public/geo/ciro_marina.geojson');
+  await fs.writeFile('public/geo/ciro_marina.json', JSON.stringify(featureCollection, null, 2));
+  console.log('Saved to public/geo/ciro_marina.json');
 }
 
 main().catch((err) => {
